Fixing code. Moving Clock and TV Guide in Schedule.js to separate files. Changing chat rooms route path. Updating Navigation links. Updating summary in chat rooms and waiting page.
