+++ front-end/src/components/Home/Clock.js
@@ -1,0 +1,40 @@
+import React, {Component} from 'react';
+
+class Clock extends Component {
+    constructor(props) {
+        super(props);
+        this.state = {
+            date: this.props.date
+        }
+    }
+
+    tick = () => {
+        this.setState({
+          date: new Date()
+        });
+    }
+
+    componentDidMount() {
+        this.timerID = setInterval(
+          () => this.tick(),
+          1000
+        );
+    }
+    
+    componentWillUnmount() {
+        clearInterval(this.timerID);
+    }
+
+    render() {
+        const {date} = this.state;
+        const options = { weekday: 'long', year: 'numeric', month: 'long', day: 'numeric' };
+        return (
+            <div>
+                <p>{date.toLocaleDateString('en-US', options)}</p>
+                <p>Time Now:{" "}{date.toLocaleTimeString()}</p>
+            </div>
+        )
+    }
+}
+
+export default Clock;
